Single entry point for pipeline config generation
